Update Snakemake pipeline to run end-to-end for one C1 chip.
The trick to merge files using glob_wildcards only works when the input
files already exist (which is why it works for merge_fastq). It also
worked when prototyping since I was doing one step at a time. However,
when trying to run a whole chip from start to end, it couldn't resolve
the rules since the files from the previous step did not yet exist.
Luckily we know that each chip has the same 96 wells, so I was able to
use expand.