More improvements to Snakefile to enable high throughput processing:
* Provide verifyBamID with 8 CPUs. It's multithreaded!
* Duplicate rules for sorting and indexing BAM files. Re-using the rules
by treating the directory as a wildcard resulted in the directory being
inserted into the name of the log file, causing the jobs to instantly
fail without producing a log file.
* Use localrules more aggressively. Slurm allows you to submit jobs from
a compute node (unlike SGE). Thus for rules that sit in the queue longer
than it takes to just run them are better run sequentially on the same
node as the snakemake process.
* Submit 500 jobs at a time to improve reduce time waiting in the queue
due to priority. To ease the burden on Slurm, also limit the number of
jobs sumbitted to 5 per second.